+++ CHANGELOG.md
@@ -5,3 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## v0.0.2 - 2020-05-30
 - Added PyPI badge 
 - Added documentation link into setup.py 
+
+## V0.1.0 - 2020-06-01
+- BKash class is deprecated
+- Added a new class. That's called BkashWebhookListener. BKash class will completely removed from v1.0.0
+- Update README and documents